src/django_backend/requirements.txt
# Core
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.5
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-allauth==0.65.0
django-crispy-forms==2.1

# Database
psycopg2-binary==2.9.11
dj-database-url==2.1.0

# Environment & Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Security
django-ratelimit==4.1.0
PyJWT==2.8.0
cryptography==42.0.5

# Async & Task Queue
celery==5.3.6
redis==5.0.1
flower==2.0.1

# API Documentation
drf-yasg==1.21.7
coreapi==2.3.3

# File Storage
Pillow==10.2.0
whitenoise==6.6.0
boto3==1.34.28
django-storages==1.14.2

# Email
sendgrid==6.11.0
django-anymail[mailgun]==10.1

# Utils
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==23.2.1

# Development
ipython==8.18.1
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Linting & Formatting
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Other
python-crontab==3.0.0
django-cleanup==8.0.0
